+++ devenv/src/lib.rs
@@ -18,7 +18,7 @@ pub use devenv_snix_backend;
 
 pub use devenv::{
     DIRENVRC, DIRENVRC_VERSION, Devenv, DevenvOptions, ProcessOptions, RunMode,
-    SecretsNeedPrompting, ShellCommand,
+    SecretsNeedPrompting, ShellCommand, format_shell_exports,
 };
 pub use devenv_tasks as tasks;